# [refactor] Performance improvement of updateStyle()