--- a/render/tests/test-createElement.js
+++ b/render/tests/test-createElement.js
@@ -46,13 +46,6 @@ o.spec("createElement", function() {
 
 		o(vnode.dom.style["--cssVar"]).equals("red")
 	})
-	o("censors cssFloat to float", function() {
-		var vnode = m("a", {style: {cssFloat: "left"}})
-
-		render(root, vnode)
-
-		o(vnode.dom.style.float).equals("left")
-	})
 	o("creates children", function() {
 		var vnode = m("div", m("a"), m("b"))
 		render(root, vnode)