@dead-claudia 2024-11-03
@kfule Can you run the benchmarks in the browser using the internal test page in `/performance/index.html`? The cost of DOM property updates in the browser is considerably higher (between FFI overhead and the fact every string is copied in every browser before being processed further), so I'd like a better idea of how this plays out in the real world.

You can spin up any server you'd like in the repo's root and just open that page. Benchmark results will be printed in the web console.