@dead-claudia 2024-11-03
If it works as advertised, I also want to port it over to #2982, as that is *the* benchmark it's bottlenecked on most.