@kfule 2024-11-04
@dead-claudia 
Thanks for the review. I will try to measure it in detail in the actual browser.
I just started measuring, and as you guessed, performance does not seem to improve in Firefox (x64,win) at first glance. On the other hand, it appears to be improving in Edge.
For now, I'll share the results of the measurements with Firefox(x64,win), Edge(x64,win) and Safari(mac,m2).